Gathered first ideas of thesis
- Read in Hopcroft, Motwani, Ullman about finite automata (Chapter 2)
- Read in the Thomas, Automata on Infinite Objects, 1990 paper
- Read diverse papers about Büchi complementation
- Decided to do implementation part in Ruby
- Learned basics of Ruby
- Made very simple DFA and NFA implementation in Ruby
- Idea for structure of the implementation (implementation/outline.txt)

Notes theoretical part: the reason that the subset construction and the
Büchi acceptance condition don't go together (and the determinized Büchi
automata accepts a superset of the language of the nondeterministic
version) is explained in Safra, 1988 (1) and (2) (better in (2)). The
first step of the theoretical part could be to try to modify the subset
construction so that both automata accept the same language (starting
point: don't mix accepting and non-accepting states in a subset). From
these insights, the end-goal is then to come somehow from a complement
of the determinized version to the correct, nondeterministic complement
of the initial Büchi automaton. But these later steps are not yet
totally clear to me.

Notes practical part: the idea is to have DFA and NFA classes, the NFA
class having a, or several, 'determinize' method. It creates and returns
a DFA object. This will be the core. Both classes also have 'visualize'
methods that draw the automata with Graphviz. This could all be
incorporated in a a Ruby on Rails application that, as an initial stage,
allows to create NFAs and transform them into DFAs with different subset
construction flavors. More features would then follow. But as a
baseline, the aim of the practical part is always to be useful to the
theoretical part.