Tools; temporal logics
Made a survey on finite state automata tools. The most notable ones,
with regard to functionality, implemented algorithms, etc., are JFLAP
and GOAL. Especially GOAL, because it is about omega-automata and has a
command line interface, which seems easy to use. For representing
automata it uses a simple XML format.

Had a brief look at other temporal logics that are equivalent to
omega-regular languages: ETL and QPTL. Both are extensions of LTL (PTL).
In Tsay (2007) it is shown that LTL cannot describe all the
omega-regular languages.

In Tsay (2007) it is also very nicely described why Büchi automaton
complementation is used in automata-theoretic model checking, that the
high complexity of Büchi complementation makes this impracticle,
however, and that one strategy is to specify a property as an LTL
formula, negate it, and translate it to a Büchi automaton, which will be
equivalent to the complement Büchi automaton if the property would have
been specified as a Büchi automaton. However, there are two
disadvantages of using LTL formulas instead of Büchi automata for
specifiying properties. 1) LTL is less expressive than Büchi automata,
and thus there may be properties that cannot be expressed by an LTL
formula; 2) a translation from an LTL formula to a Büchi automata may
not yield an optimal (i.e. smallest) Büchi automaton. Shortly put, it
would be most desirable to be able to specify properties as Büchi
automata and then simply complement them, if only the complementation
wouldn't be so computationally complex...