Started translation from YAML to GOAL File Format
The idea is to connect GOAL to our tool and using it for equivalence tests
between automata and possibly other operations. A translation from YAML
to the GOAL file format and vice versa makes this possible. Started with
function yaml_to_goal.

Changed the representation of transitions in the YAML file to a list
with items of the form "p -> q: 0,1". This is completed.

Changed the internal representation of transitions from
{from=>{symbol=>{to1,to2},..},..} to {from=>{to=>{sym1,sym2},..},..}.
This is not yet completed.