How to install and use Graphviz on Heroku
Graphviz can be installed on Heroku by the means of Heroku buildpacks.
To use Graphviz with a Rails app, the two buildpacks
heroku-buildpack-ruby and heroku-buildpack-graphviz must be used. To use
multiple buildpacks for an app, there is the buildpack
heroku-buildpack-multi, which will be the real buildpack of the app (the
other buildpacks are specified in the file .buildpacks in the root
directory).

Deploying the app to Heroku (git push heroku master) will install Ruby
and Graphviz on Heroku. The installation of Graphviz on Heroku worked so
far, but I have not yet tested if using Graphviz from the app works.

There's another app that uses Graphviz on Heroku (in Python):
https://github.com/pygram/pygram. Another website,
http://hackingoff.com/compilers/nfa-to-dfa-conversion, probably not on
Heroku uses Graphviz as well, in a similar way that our project is
intended to use it.

For installing and using GOAL on Heroku, it would be necessary to create
a new buildpack that fetches and installs GOAL on Heroku. It remains to
be seen if this is feasible.