Graphviz on Heroku and buildpacks
Took a closer look at the environment of an app on Heroku and how
buidpacks work. The heroku-buildpack-graphviz seems to be very flawed
and I will write a new one. Also looked at how the
heroku-buildpack-multi works.

Installed Graphviz manually on Heroku (via heroku run bash). It
principially works but there seem to be fonts missing. It should be
possible to fix it, however.