Ways for instaling Graphviz on Heroku
There are two, almost three, ways to install Graphviz on Heroku.
1. Compile and install from sources
2. Install the graphviz Debian package with apt-get
3. (Install the graphviz Debian package "manually" with dpkg)

1 has the problem that Graphviz is not fully functional due to
dependencies which are difficult to resolve. More precisely, png and pdf
output format don't work. The reason is that apparently certain
libraries are not found or missing. Among the working output formats are
svg and eps. An advantage of this method is that any Graphviz version
can be chosen.

In 2, Graphviz seems to be fully functional, all the output formats
work. A disadvantage is that the Graphviz version in the only accessible
Debian package is very old (2.20.13). Another disadvantage is that the
default font of the output graphs is some sans serif font. I don't know
if this is intentionally or not.

3 didn't work. On graphviz.org there are .deb packages of all Graphviz
versions. The idea was to fetch the desired version from there and
install it with dpkg. I also tried to install all the dependency
packages. However, when executing Graphviz, there's an error due to
missing or not matching dependencies.

I will create buildpacks for 1 and 2. This way, the graph rendering
characteristics of both approaches can be evaluated and the better one
chosen.