Buildpack does not yet work
For some reason, everything done by my test buildpack has no effect.
That is, after the build none of the changes made during the build are
there. Next step is to find out what the cause is. Otherwise, it is
quite clear how buildpacks work, and it is acutally also quite easy to
write them.

The installation of GOAL on Heroku works and is very easy. GOAL is a
Java program and Java is already installed on Heroku. It is enough to
copy the GOAL program directory to Heroku and the program can instantly
be used.