Prototype of web application, plugin for GOAL
Finished a first prototype of the web application
(https://github.com/weibeld/buechi). Everything works, just invoking a
GOAL command is very slow. Nailgun seems to not work correctly on Heroku
(locally it works well, in combination with the web server).

Looked at how to write plugins for GOAL. The idea is to integrate our
complementation algorithms in GOAL. Currently writing shell scripts for
generating and compiling plugins. Deploying is then very easy, just copy
the directory containing the plugin to the plugins/ directory of GOAL.