Cleaned up
Cleaned repository up. Integrated the buechi_transformer Rails project,
which is itself a Git repository, in this repository by simply copying
it to this repository.

This way, the outer repository contains all the files and file changes,
including the ones in the inner repository. The inner repository is
completely untouched by the outer repository. This is exactly what I
need.