Nailgun script (working)
Adapted the script of the web app for executing GOAL command line
commands with Nailgun. Until today, it didn' work for GOAL commands that
take a GFF file as argument. Today it suddenly worked. I don't know what
the change is that made it work.

The execution time of GOAL command line commands can be reduced by 2-2.5
seconds. However, it doesn't affect the startup time of the GOAL GUI by
clicking on the application icon or on a GFF file.

One thing that doesn't work with Nailgun enabled is to start the GOAL
GUI from the command line (running 'goal'). The GOAL GUI can be started
and used, but when one tries to close it, it doesn't respond to
anything.