Finished implementation of entire algorithm
Implemented the second stage of the construction. The optimisation of
deleting states whose rightmost component has colour 2 is fixedly
included. Completing the input automaton (if it is not already complete)
is also fixedly included.

Next steps are to improve the code, and make the optimisations and other
options accessible as options, which are checkboxes in the GUI (after
clicking on the algorithm's menu item) and command line options in the
command line mode.