Modifications on scripts
In plugin_build.sh, removed the need to specify an rt.jar for the
-Xbootclasspath option to javac. It works without that option, it just
produces a compiler warning, but the code works as far as I have seen.
Tested the script on Linux as well.

Made complement.sh Linux compatible. Some commands had to be changed for
Linux. Tested it on Linux and it works well.

Sent the plugin and these two scripts to Ulrich.