Added options
Ming-Hsien responded, and with his instruction I could add the options
to the construction, both, in the GUI and on the command line. Up to
now, two options, MakeComplete and IgnoreRightColor2. It works well so
far, just needs adjusting of some details. Sent another e-mail with two
questions to Ming-Hsien.

Next steps are:
* Finishing the options business
	* Help text on command line (asked Ming-Hsien)
	* Saving selection of options in the GUI (asked Ming-Hsien)
	* Organise code better
* Add support for propositional alphabets
	* ...by translating a propositional one to a classical one
	  before, and translating it back after the construction
* Modify the complement.sh script so that an algorithm and its options
  don't need to be enclosed in quotes anymore