Finished the options business
Added help text on command line, enabled persistent options preferences
in the GUI, and improved organisation of code.

Each option has a default value that is persistently set in the
preferences file. For specifying options on the command line, the idea
is that for every option that is not specified on the command line, the
default value apply. For options that are specified, the inverse of
the default value applies (probably all options will be boolean). Like
this, options that are always used anyway can be assigned a default
value of true, and then they don't need to be specified on the command
line each time. Changing the default values then however requires
adapting the option descriptions in the help text, and probably the
option token, like a '+' prefix for enabling and a '-' prefix for
disabling something. However, this could in turn become complicated for
the user if certain combinations of options are not allowed.

Encountered the question of how to make the deterministic interim
automaton (output of first stage) complete if it isn't.

Next steps are thus (in the following order of importance):
1) How to make the interim automaton complete. Implement it.
2) Add handling of the propositional alphabet
3) Modify complement.sh so that no quotes are needed for the algorithm