Propositional alphabet and complement.sh
Added the handling of automata with a propositional alphabet. Such
automata are converted to classical ones before the construction, then
the construction is done with their classical equivalent, and after the
construction the classical representation is converted back to a
propositional one.

Modified complement.sh so that all the command line arguments after the
second argument and starting with a '-' are taken as algorithm options.
No quotes are needed anymore.

The last TODO is now still the issue with the completeness of the
deterministic interim automaton if the input automaton is not made
complete at the beginning.