Implemented handling of incomplete automata
If a state is incomplete (i.e. no outgoing transitions for one or more
symbols), this fact is at first ignored during the two stages of the
construction, and after the second stage a sink state is added to the
automaton with all the missing transitions from the upper part of the
automaton. The sink state is accepting.