First draft of colour 2 reduction optimisation
Could make the implementation run, after a very long debugging. The
reason was that I didn't test for an empty successor set for a component
of state p in two places. This resulted in empty components added to the
states what provoked infinitely many distinct states, and this caused an
infinite loop, I think.

The program runs, but there are still many errors in the implementation
of the optimisation.