UBELIX cluster access, installation of libefa
Could install libefa on Debian (virtual machine) and on Mac. On Mac it
doesn't seem to work correctly, and it will probably never do. It has
been written only with Linux in mind. On Debian it seems to work.

Got an account for the Linux cluster of the University of Bern (UBELIX).
The cluster looks very nice.

Running GOAL on the cluster should be no problem. However libefa is most
probably not possible. I cannot even install it on the submit host
because the tools needed for the installation are not present. Even if I
could install it on the submit host, it would probably not work on the
hosts in the cluster where it will be executed. If I want to use libefa,
I have to run it on Debian in the virtual machine, or an a Linux host
that I have access to.