Zeittotschlagen im Militär
Tried to add a step-by-step version of the construction. Asked
Ming-Hsien for more information.

Changed the merging behaviour to the correct one, i.e. []() -> []
instead of ()[] -> []. The previous one was clearly wrong what can be
seen with the complement_equiv.gs script. With the new merging rule it
seems to be correct, at least the first few hundred automata of
complement_equiv.gs didn't yield a counterexample.