GOAL script, UBELIX cluster
Made a GOAL script for checking the correctness of the complementation
construction. Generates random automaton, complements it with an
existing algorithm, complements it with the algorithm to test, and
compares the two complements for equivalence. It's in fact the same as
the existing script complement_equiv.gs.

Figured out how to execute jobs on UBELIX. The cluster can be accessed
from outside the UNIBE network. It works without problems from at home.