New version of GOAL, UBELIX setup
Installed and tested new version of GOAL (2014-08-08). There are some
differences, including the behaviour when trying to pass options to a
complementation construction in a variable in GOAL scripts. Asked
Ming-Hsien for more information.

Set up a job management environment on the submit node of UBELIX. It
includes a script (submit.sh) which allows to easily submit jobs with
default options and optionally specify options that then overwrite the
default options. Each job has its own directory and all the produced
files are located in this directory.