Fixed problems from yesterday; improved scripts
The strange problems with the execution of GOAL were due to multiple
jobs accessing the GOAL program files in the GPFS home directory.
Resolved this by making the jobs copy the whole GOAL folder to their
local scratch.

Several improvements to the scripts and the cluster environment:
- Algorithm and options can now be passed as command line arguments to
  complement.sh
- There's only one complement.sh in a central location that is
  referenced by all qsub exections. The qsub execution directory can be
specified explicitly to submit.sh and is not automatically the script
directory anymore.
- Of the data and the GOAL directory, archives are maintained that can
  be directly copied by the jobs to their local scratch

Started an experiment on hnodes through the mpi.q. Complementation of
the 11'000 size 15 automata. Tested algorithms are "fribourg", "fribourg
-r2ifc -m", and "piterman -eq -macc -sp -ro -r".