Peculiarities of -macc and -r options; added -rr option
The -macc option very often causes larger complements which is
counterintuitive. In other algorithms (piterman) this is the case too,
but less often. I don't know the reason for this. The real degree of
this will become apparent when the experiments with the full test set.

The -r option removes only unreachable and dead state from the output
automaton. It doesn't remove states from the input automaton. This is
identical to the -r option of the other algorithms. If we have the
automata 'in' and 'in_red' ('in' with the unreachable and dead states
removed), then the other algorithms with the -r option yield different
results for 'in' and 'in_red'. That means, the -r option does not reduce
the input automaton first. This is just here because Joel initially
believed that the -r option would also remove the unreachable and dead
states from the input automaton.

Added an explicit option -rr for removing the unreachable and dead
states from the input automaton. Since the other algorithms don't do
this, this option must not be enabled during the comparative
experiments. The automata in the test set contain unreachable and dead
states, so reducing the input automata with the -rr option while the
other algorithms don't do this would probably make a huge difference.