Issue with -m2 not yet resolved
For all the automata in bad_m2_examples, the construction is now correct
with the modification made in FribourgConstruction.java. The
modification consists in making both, the acc and nonacc successors of
the current M2 colour 2. Consquently, they are then merged to one
component.

However, this doesn't resolve the issue for all automata. I run another
complement-equivalence experiment and there are still some automata for
which the construction is incorrect (18 out of 678, i.e. 2.6 percent).