Plugin -m and -m2 options rewrite
Started to rewrite the implementation of the -m and -m2 options. -m is
done, compiles but not yet tested. -m2 is not yet done. The idea is to
rewrite it in a CORRECT way. Currently the idea is to remember the
components which would be 2-coloured without -m2 but are just 1-coloured
because of -m2, and then when a *-component disappears to treat them
differently. I don't know however if this is correct. If not, the code
is at least more flexible (and better) to implement a correct version of
the -m option.