Rewrite of colour determination part; started rewrite of -m2
Rewrote the part of the construction that determines the colour of the
successor components in q of all the components in p. Started a
complement-equivalence test to see if the construction is still correct.

Started to rewrite the implementation of the -m2 option. Maybe it's
nevertheless best to implement exactly the version of Joel and Ulrich,
instead of trying an own version with the "suppressed colour 2" idea.
The code is now relatively comprehensible and flexible, and it should be
possible to add the -m2 option from scratch relatively easily.