New version of the -m2 option
Implemented the -m2 option following the version of Joel. It's not yet
correct and I guess the current state is more or less the same as with
the old implementation. The bad_m2_examples are incorrect and can be
made correct by making both childs of the sole component that may have
2-coloured successors colour 2. Currently running a test to see how many
automata out of 1000 are still incorrect (with this little fix). With
the old implementation, 18 automata out of 678 were still incorrect.

So, the new implementation as it is now seems to be quite similar to the
old one. But what I want to try now is the newest version from Joel in
which, when a 2-coloured component disappears, a 1-coloured component in
the very next state is made 2-coloured and not only in the state after
this next state. This next state must then still be accepting (even
though it contains a 2-coloured component), and to achieve this, it can
be marked with a flag (e.g. a *).