Meeting with Joel
Met Joel in the afternoon. The following points have been clarified
(they all concern the version of -m2 where a () is directly made [] in a
state where a [] is emptied and the state is marked as accepting):

1. Both children of a [], or of a () in a state with no [] and where
there have been no [] successors so far, have to be []. Making the
accepting set [] and the non-accepting set () is not correct.

2. The component merging can be done before transforming a () to a [],
but then it has to be done after that again, because, first, the empty
[] which influenced the merging is away, and second, the colour of one
component changed, what may result in further merging options).

3. An empty [] blocking a merging (if the merging is done before the
transformation of a () to a []) should not be a problem. Also, doing the
merging before or after should leave the algorithm correct (the output
automata may however be different).

4. A state where a [] has been emptied and that contains no {} can be
removed. The reason is that from this state no accepting state can be
reached infinitely often.

5. The procedure for selecting a () to make [] may be the crucial point.
However, taking the first () after the first {} to the left of the
emptied [] should be correct. If there is no {}, then, as mentioned, the
state can be removed. If there is no (), then nothing is done. So, there
is always at least one {} and one () when this procedure is applied, and
then it's quite clear which () to select.