-m2 works!
The version of the code in rewrite_3 now works. I have edited this
version already before the last commit and adapted it accurately to the
points described there. I did also some refactoring and improvements on
the code. Then I did complementation-equivalence tests with fribourg -m
-m2 on totally 5000 automata, and all are correct. Also the automata
which have been done incorrectly in the previous tests (I collected
them) are now done correctly.

I don't know yet what exactly it was that made it working. I can
investigate this later. For now, most important is that the plugin
version in rewrite_3 is the latest and working version.