Ported code to new GOAL version (2014-10-16)
Currently, there are two versions of the code, one for the old version
(2013-07-11) and one for the new (2014-10-16). The code seems to work
well for the new version. The intention is to fully go over to this new
version as soon as it is officially released. Then also do the
experiments etc. with it.