New GOAL version
Ming-Hsien did the following changes on the previous version
(2014-10-16):
- Fixed a problem with the option passing in piterman
- Allow accessing shell variables in GOAL scripts
- Fixed error with the list of the command line complementation options
  in code
He plans to release the new version on October 25, 2014. By the way, the
other changes of this version, compared to the 2013-07-11 version are:
- No growth of message panel in step-by-step complemenation
- Allow complementation options to be passed as variables in GOAL
  scripts
- Shortcut for complementation: Ctrl-Alt-C

In plugin code:
- Added second constructor to FribourgConstruction which allows the
  construction to be used for the containment and equivalence operations
The plugin is now really fully integrated in GOAL. It could be the final
state of the code now.