Analysis of Fribourg construction; run analysis
Analysed the Fribourg construction again. Compared with the slice-based
construction in Vardi, Wilke (2007). Seems to be very similar.

Maybe have to insert a section in Chapter 2 about run analyis, and
present run trees, split trees, reduced split trees, and run DAGs. The
Fribourg construction and the slice-based construction is based on
reduced split trees (introduced by Muller, Schupp, 1995), and the
rank-based constructions are based on run DAGs. With the run trees and a
good example it becomes clear why the subset-construction does not work
for determinising NBW.