TikZ framework for drawing automata, run trees, etc.
Created drafts of the figures showing an example automata and a
corresponding run tree, split tree, reduced split tree, and run DAG. To
this end, created kind of a framework for having a uniform style for
automata and other structures containing states and transitions.

Received the fulltext of "Improved Ramsey-Based Büchi Complementation"
from Stefan Breuers.

The slice-based construction described in "Automata: From Logics to
Algorithms" has large parts that are similar to the Fribourg
construction, but it is still different. Joel was not aware of this
paper, and they came to similar formalisations for the Fribourg
construction independently.