Starting to gather results
Run some more jobs. Starting to gather and complete results and add them
to the report. Idea of adding a command line command to GOAL for testing
if an automaton is complete. This is for the analysis of the test set to
see how many and which automata are complete.