R functions for plotting 'persp' and 'contour'; all results
All the tests have successfully finished (the 10 'rank' jobs on the 10
partitions of the GOAL testset almost used the full 72 hours). The results
are copied results/.

Various R functions for analysing the results, including drawing 'persp' and
'contour' plots. The idea is then to write small R programs which produce the
desired output graphics.